+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@cfn-modules/asg-singleton-amazon-linux2",
-  "version": "2.2.0",
+  "version": "2.3.0",
   "description": "AWS Auto Scaling Group running a single EC2 instance based on Amazon Linux 2 with a fixed public IP address (Elastic IP), alerting, IAM user SSH access, following an immutable infrastructure approach (root volume can be replaced at any time)",
   "author": "Michael Wittig <michael@widdix.de>",
   "license": "Apache-2.0",
