+++ test/with-hosted-zone-private.yml
@@ -25,4 +25,5 @@ Resources:
         HostedZoneModule: !GetAtt 'HostedZone.Outputs.StackName'
         SubnetReach: Private
         AmazonLinux2Version: '2.0.20181114'
+        InstanceType: 't3.nano'
       TemplateURL: './node_modules/@cfn-modules/asg-singleton-amazon-linux2/module.yml'
