+++ test/with-hosted-zone-public.yml
@@ -24,4 +24,5 @@ Resources:
         VpcModule: !GetAtt 'Vpc.Outputs.StackName'
         HostedZoneModule: !GetAtt 'HostedZone.Outputs.StackName'
         AmazonLinux2Version: '2.0.20181114'
+        InstanceType: 't3.nano'
       TemplateURL: './node_modules/@cfn-modules/asg-singleton-amazon-linux2/module.yml'
