+++ test/with-key-name.yml
@@ -22,6 +22,7 @@ Resources:
         VpcModule: !GetAtt 'Vpc.Outputs.StackName'
         KeyName: !Ref KeyName
         AmazonLinux2Version: '2.0.20181114'
+        InstanceType: 't3.nano'
       TemplateURL: './node_modules/@cfn-modules/asg-singleton-amazon-linux2/module.yml'
 Outputs:
   PublicIpAddress:
